Fix DDB table Amazon Owned KMS key with replicas
Some time ago Amazon changed the default DDB encryption behaviour to use
a key 'Managed by DynamoDB' (it lives in the DDB internal account) if
setting `server_side_encryption.enabled = false`

The current update table logic, if using replicas, assumes you are using
a CMK. If this is not the case, an update table request is triggered
with an empty `kMSMasterKeyId`[1] and AWS API throws the below error.

This commit fixes the logic to only update replicas if there is a CMK
used.

[1]
```
│ Error: updating DynamoDB Table (BADGERS) SSE: ValidationException: 4 validation errors detected: Value '' at 'replicaUpdates.1.member.update.kMSMasterKeyId' failed to satisfy constraint: Member must have length greater than or equal to 1; Value '' at 'replicaUpdates.2.member.update.kMSMasterKeyId' failed to satisfy constraint: Member must have length greater than or equal to 1; Value '' at 'replicaUpdates.3.member.update.kMSMasterKeyId' failed to satisfy constraint: Member must have length greater than or equal to 1; Value '' at 'replicaUpdates.4.member.update.kMSMasterKeyId' failed to satisfy constraint: Member must have length greater than or equal to 1
    ```

Fixes https://github.com/hashicorp/terraform-provider-aws/issues/31153