feat: Add required suffix when specifying log group ARN in FIS experiment template
Specifying a cloudwatch log-group ARN on a `aws_fis_experiment_template` like so:
```
  log_configuration {
    log_schema_version = 2

    cloudwatch_logs_configuration {
      log_group_arn = aws_cloudwatch_log_group.example.arn
    }
  }
```
results in:
```
╷
│ Error: creating AWS FIS (Fault Injection Simulator) Experiment Template (Test dev provider): operation error fis: CreateExperimentTemplate, https response error StatusCode: 400,
| RequestID: a84b03ea-b1ff-49f5-9be6-b0ad450666cf, ValidationException: The logGroupArn "arn:aws:logs:eu-west-1:${accountId}:log-group:/aws/fis" in cloudWatchLogsLogConfiguration is not valid.
│
│   with aws_fis_experiment_template.fis,
│   on fis.tf line 1, in resource "aws_fis_experiment_template" "fis":
│    1: resource "aws_fis_experiment_template" "fis" {
│
╵
```
This is because AWS expects this suffix `:*` on the log-group ARN like so:
```
  log_configuration {
    log_schema_version = 2

    cloudwatch_logs_configuration {
      log_group_arn = "${aws_cloudwatch_log_group.example.arn}:*"
    }
  }
```

As far as I could tell, the requirement for the suffix is not well documented. Also the ambiguous Api error from AWS is not helpful.
Propose the required suffix is added if not present. Specifying `log_group_arn = aws_cloudwatch_log_group.example.arn` is more
intuitive imho.